+++ packages/headless/esbuild.mjs
@@ -41,6 +41,7 @@ const quanticUseCaseEntries = {
   recommendation: 'src/recommendation.index.ts',
   'case-assist': 'src/case-assist.index.ts',
   insight: 'src/insight.index.ts',
+  commerce: 'src/commerce.index.ts',
 };
 
 function getUmdGlobalName(useCase) {
@@ -160,6 +161,24 @@ const browserUmd = Object.entries(useCaseEntries).map((entry) => {
   );
 });
 
+const codeReplacerPlugin = (filePath, target, replacement) => ({
+  name: 'code-replacer-plugin',
+  setup(build) {
+    build.onLoad({filter: /\.ts$/}, async (args) => {
+      if (args.path.endsWith(filePath)) {
+        const fs = require('fs');
+        const source = await fs.promises.readFile(args.path, 'utf8');
+        const modifiedSource = source.replace(target, replacement);
+        return {
+          contents: modifiedSource,
+          loader: 'ts',
+        };
+      }
+      return undefined;
+    });
+  },
+});
+
 const quanticUmd = Object.entries(quanticUseCaseEntries).map((entry) => {
   const [useCase, entryPoint] = entry;
   const outDir = getUseCaseDir('dist/quantic/', useCase);
@@ -181,7 +200,14 @@ const quanticUmd = Object.entries(quanticUseCaseEntries).map((entry) => {
         'ponyfills/abortable-fetch-shim.js',
         '../../node_modules/navigator.sendbeacon/dist/navigator.sendbeacon.cjs.js',
       ],
-      plugins: [umdWrapper({libraryName: globalName})],
+      plugins: [
+        umdWrapper({libraryName: globalName}),
+        codeReplacerPlugin(
+          'src/api/knowledge/answer-slice.ts',
+          '(updatedArgs, api, extraOptions)',
+          '(updatedArgs,{...api, signal: null},extraOptions)'
+        ),
+      ],
     },
     outDir
   );
@@ -269,6 +295,7 @@ const nodeEsm = Object.entries(useCaseEntries).map((entry) => {
     {
       entryPoints: [entryPoint],
       outfile,
+      sourcemap: true,
       format: 'esm',
       external: ['pino'],
       mainFields: ['module', 'main'],