+++ packages/headless/src/controllers/commerce/core/parameter-manager/headless-core-parameter-manager.ts
@@ -1,9 +1,15 @@
 import {RecordValue, Schema, SchemaDefinition} from '@coveo/bueno';
-import {UnknownAction} from '@reduxjs/toolkit';
-import {CommerceEngine} from '../../../../app/commerce-engine/commerce-engine.js';
+import {createSelector, UnknownAction} from '@reduxjs/toolkit';
+import {
+  CommerceEngine,
+  CommerceEngineState,
+} from '../../../../app/commerce-engine/commerce-engine.js';
 import {stateKey} from '../../../../app/state-key.js';
 import {Parameters} from '../../../../features/commerce/parameters/parameters-actions.js';
+import {parametersReducer as commerceParameters} from '../../../../features/commerce/parameters/parameters-slice.js';
+import {CommerceParametersSection} from '../../../../state/state-sections.js';
 import {deepEqualAnyOrder} from '../../../../utils/compare-utils.js';
+import {loadReducerError} from '../../../../utils/errors.js';
 import {validateInitialState} from '../../../../utils/validate-payload.js';
 import {
   Controller,
@@ -15,7 +21,15 @@ export interface ParameterManagerProps<T extends Parameters> {
   /**
    * The initial state that should be applied to the `ParameterManager` sub-controller.
    */
-  initialState: ParameterManagerInitialState<T>;
+  initialState?: ParameterManagerInitialState<T>;
+
+  /**
+   * Whether the controller's state should exclude the default parameters returned by the Commerce API, and only include
+   * the parameters that were set explicitly set through dispatched actions.
+   *
+   * Defaults to `false`.
+   */
+  excludeDefaultParameters?: boolean;
 }
 
 export interface CoreParameterManagerProps<T extends Parameters>
@@ -39,16 +53,6 @@ export interface CoreParameterManagerProps<T extends Parameters>
    * The action to dispatch to fetch more results.
    */
   fetchProductsActionCreator: FetchProductsActionCreator;
-
-  /**
-   * Enriches the parameters with the active parameters.
-   * @param state
-   * @param activeParams
-   */
-  enrichParameters(
-    state: CommerceEngine[typeof stateKey],
-    activeParams: T
-  ): Required<T>;
 }
 
 export interface ParameterManagerInitialState<T> {
@@ -113,26 +117,40 @@ export function buildCoreParameterManager<T extends Parameters>(
   engine: CommerceEngine,
   props: CoreParameterManagerProps<T>
 ): ParameterManager<T> {
+  if (props.excludeDefaultParameters && !loadParameterManagerReducers(engine)) {
+    throw loadReducerError;
+  }
+
+  const parametersSelector = createSelector(
+    (state: CommerceEngineState) => state.commerceParameters,
+    (parameters) => parameters
+  );
+
   const {dispatch} = engine;
   const controller = buildController(engine);
 
-  validateInitialState(
-    engine,
-    initialStateSchema(props.parametersDefinition),
-    props.initialState,
-    'buildCoreParameterManager'
-  );
-  dispatch(props.restoreActionCreator(props.initialState.parameters));
+  if (props.initialState) {
+    validateInitialState(
+      engine,
+      initialStateSchema(props.parametersDefinition),
+      props.initialState,
+      'buildCoreParameterManager'
+    );
+    dispatch(props.restoreActionCreator(props.initialState.parameters));
+  }
 
   return {
     ...controller,
 
     synchronize(parameters: T) {
       const activeParams = props.activeParametersSelector(engine[stateKey]);
-      const oldParams = props.enrichParameters(engine[stateKey], activeParams);
-      const newParams = props.enrichParameters(engine[stateKey], parameters);
 
-      if (deepEqualAnyOrder(oldParams, newParams)) {
+      // Always restore empty parameters or parameters that are different from the active ones.
+      // We always restore empty parameters in commerce because they may correspond to navigation to a new page.
+      if (
+        Object.keys(parameters).length > 0 &&
+        deepEqualAnyOrder(activeParams, parameters)
+      ) {
         return;
       }
 
@@ -141,8 +159,20 @@ export function buildCoreParameterManager<T extends Parameters>(
     },
 
     get state() {
-      const parameters = props.activeParametersSelector(engine[stateKey]);
-      return {parameters};
+      return {
+        parameters: props.excludeDefaultParameters
+          ? (parametersSelector(engine[stateKey]) as T)
+          : props.activeParametersSelector(engine[stateKey]),
+      };
     },
   };
+}
+
+function loadParameterManagerReducers(
+  engine: CommerceEngine
+): engine is CommerceEngine<CommerceParametersSection> {
+  engine.addReducers({
+    commerceParameters,
+  });
+  return true;
 }